requirements.txt
absl-py==2.3.0
accelerate==1.7.0
addict==2.4.0
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
airportsdata==20250622
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
altair==5.5.0
annotated-types==0.7.0
antlr4-python3-runtime==4.7.2
anyio==4.9.0
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
arxiv==2.2.0
astor==0.8.1
asttokens==3.0.0
async-lru==2.0.5
async-timeout==5.0.1
attrdict==2.0.1
attrs==25.3.0
av==14.4.0
babel==2.17.0
beautifulsoup4==4.13.4
binpacking==1.5.2
bitsandbytes==0.46.0
blake3==1.0.5
bleach==6.2.0
blessed==1.21.0
blinker==1.9.0
Brotli==1.1.0
cachetools==6.1.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
comm==0.2.2
compressed-tensors==0.9.3
contourpy==1.3.0
cos_python_sdk_v5==1.9.37
cpm-kernels==1.0.11
crcmod==1.7
cryptography==45.0.4
cupy-cuda12x==13.4.1
cycler==0.12.1
dacite==1.9.2
datasets==3.2.0
debugpy==1.8.14
decorator==5.2.1
decord==0.6.0
deepspeed==0.17.1
defusedxml==0.7.1
Deprecated==1.2.18
depyf==0.18.0
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
dnspython==2.7.0
duckduckgo_search==5.3.1b1
einops==0.8.1
email_validator==2.2.0
et_xmlfile==2.0.0
evalscope==0.16.0
evaluate==0.4.4
exceptiongroup==1.3.0
executing==2.2.0
fastapi==0.115.13
fastapi-cli==0.0.7
fastjsonschema==2.21.1
fastrlock==0.8.3
feedparser==6.0.11
ffmpy==0.6.0
filelock==3.18.0
fire==0.7.0
fonttools==4.58.4
fqdn==1.5.1
frozenlist==1.7.0
fsspec==2024.2.0
func_timeout==4.3.5
future==1.0.0
fuzzywuzzy==0.18.0
gguf==0.17.1
gitdb==4.0.12
GitPython==3.1.44
googleapis-common-protos==1.70.0
gpustat==1.1.1
gradio==4.44.1
gradio_client==1.3.0
griffe==1.7.3
grpcio==1.73.0
h11==0.16.0
h2==4.2.0
h5py==3.14.0
hf-xet==1.1.3
hjson==3.1.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.33.0
human-eval==1.0.3
hyperframe==6.1.0
idna==3.10
imageio==2.37.0
immutabledict==4.2.1
importlib_metadata==8.0.0
importlib_resources==6.5.2
interegular==0.3.3
ipykernel==6.29.5
ipython==8.18.1
ipywidgets==8.1.7
isoduration==20.11.0
jedi==0.19.2
jieba==0.42.1
Jinja2==3.1.6
jiter==0.10.0
jmespath==0.10.0
joblib==1.5.1
json5==0.12.0
jsonlines==4.0.0
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.3
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
kiwisolver==1.4.7
lagent==0.2.3
langdetect==1.0.9
lark==1.2.2
latex2sympy2==1.9.1
lazy_loader==0.4
Levenshtein==0.27.1
llguidance==0.7.30
llvmlite==0.43.0
lm-format-enforcer==0.10.11
lxml==5.4.0
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.4
matplotlib-inline==0.1.7
mdurl==0.1.2
mistral_common==1.6.2
mistune==3.1.3
mmengine==0.10.7
mmengine-lite==0.10.7
modelscope==1.27.1
mpmath==1.3.0
ms-opencompass==0.1.6
ms-vlmeval==0.0.17
ms_swift==3.5.2
msgpack==1.1.1
msgspec==0.19.0
multidict==6.5.0
multiprocess==0.70.16
narwhals==1.43.1
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
ninja==1.11.1.4
nltk==3.9.1
notebook==7.4.3
notebook_shim==0.2.4
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu11==11.11.3.6
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu11==8.7.0.84
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu11==10.3.0.86
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu11==11.7.5.86
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-ml-py==12.575.51
nvidia-nccl-cu11==2.20.5
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu11==11.8.86
nvidia-nvtx-cu12==12.4.127
omegaconf==2.0.0
openai==1.90.0
OpenCC==1.1.9
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
opentelemetry-api==1.26.0
opentelemetry-exporter-otlp==1.26.0
opentelemetry-exporter-otlp-proto-common==1.26.0
opentelemetry-exporter-otlp-proto-grpc==1.26.0
opentelemetry-exporter-otlp-proto-http==1.26.0
opentelemetry-proto==1.26.0
opentelemetry-sdk==1.26.0
opentelemetry-semantic-conventions==0.47b0
opentelemetry-semantic-conventions-ai==0.4.9
orjson==3.10.18
oss2==2.19.1
outlines==0.1.11
outlines_core==0.1.26
overrides==7.7.0
packaging==25.0
pandas==2.3.0
pandocfilters==1.5.1
parso==0.8.4
partial-json-parser==0.2.1.1.post6
peewee==3.18.1
peft==0.13.2
pexpect==4.9.0
phx-class-registry==4.0.6
pillow==10.4.0
platformdirs==4.3.8
portalocker==3.2.0
prettytable==3.16.0
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.22.1
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==4.25.8
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==20.0.0
pyarrow-hotfix==0.7
pycountry==24.6.1
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
pydub==0.25.1
Pygments==2.19.1
pynvml==12.0.0
pyparsing==3.2.3
pypinyin==0.54.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-json-logger==3.3.0
python-Levenshtein==0.27.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.0
qwen-vl-utils==0.0.8
rank-bm25==0.2.2
RapidFuzz==3.13.0
ray==2.47.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.0.0
rich-toolkit==0.14.7
rouge==1.0.1
rouge-chinese==1.0.3
rouge_score==0.1.2
rpds-py==0.25.1
ruff==0.12.0
sacrebleu==2.5.1
safetensors==0.5.3
scikit-image==0.24.0
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
semantic-version==2.10.0
Send2Trash==1.8.3
sentence-transformers==4.1.0
sentencepiece==0.2.0
sgmllib3k==1.0.0
shellingham==1.5.4
simplejson==3.20.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
socksio==1.0.0
sortedcontainers==2.4.0
soupsieve==2.7
stack-data==0.6.3
starlette==0.46.2
streamlit==1.46.0
sty==1.0.6
swanboard==0.1.7b1
swankit==0.1.5
swanlab==0.4.12
sympy==1.13.1
tabulate==0.9.0
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
termcolor==3.1.0
terminado==0.18.1
threadpoolctl==3.6.0
tifffile==2024.8.30
tiktoken==0.9.0
timeout-decorator==0.5.0
tinycss2==1.4.0
tokenizers==0.21.1
toml==0.10.2
tomli==2.2.1
tomlkit==0.12.0
torch==2.6.0
torch-geometric==2.6.1
torch_cluster==1.6.3
torch_scatter==2.1.2
torchaudio==2.6.0
torchvision==0.21.0
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.51.3
transformers-stream-generator==0.0.5
triton==3.2.0
trl==0.19.0
typer==0.16.0
types-python-dateutil==2.9.0.20250516
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
ujson==5.10.0
uri-template==1.3.0
urllib3==2.4.0
uvicorn==0.34.3
uvloop==0.21.0
validators==0.35.0
vllm==0.8.5.post1
watchdog==6.0.0
watchfiles==1.1.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.1.3
widgetsnbextension==4.0.14
word2number==1.1
wrapt==1.17.2
xformers==0.0.29.post2
xgrammar==0.1.18
xlsxwriter==3.2.5
xmltodict==0.14.2
xtuner==0.1.23
xxhash==3.5.0
yapf==0.43.0
yarl==1.20.1
zipp==3.23.0
zstandard==0.23.0
